OAK-9238 oak-run explore should support Azure Segment Store
Added AbstractSegmentTarExplorerBackend in order to share common code between SegmentTar and Azure Segment Store implementation.
